refactor(frontend): extract contacts filter empty state into its own component
Pulls the ~30 lines of empty-state markup out of
`TransactionsFilterContactsPanel` and into a new
`TransactionsFilterContactsEmptyState.svelte`. The panel now only
branches on `$contacts.length === 0` and delegates the empty-state UI
(illustration, "no contacts yet" line, OISY-protects-you lockup,
Learn more link, "Add your first contact" CTA, `openAddressBook`
handler) to the new component.

The panel's import list drops `IconAddressBook`, `IconShieldCheck`,
`Button`, the `TRANSACTIONS_FILTER_CONTACTS_EMPTY_CTA` test id,
`modalStore`, and `replaceOisyPlaceholders`. The panel keeps the
list/search/tracking logic that's actually about filtering the
non-empty contact list.

Tests follow the same shape: the new component owns the empty-state
spec (title, OISY lockup + description, Learn more link, CTA opens
the address book), and the panel spec keeps a single branching
assertion that the empty-state component is rendered when contacts
is empty.

Addresses AntonioVentilii's review comment on PR #12778.
