+++ src/frontend/src/lib/components/transactions/filter/TransactionsFilterContactsPanel.svelte
@@ -1,11 +1,8 @@
 <script lang="ts">
 	import { Checkbox } from '@dfinity/gix-components';
 	import Avatar from '$lib/components/contact/Avatar.svelte';
-	import IconAddressBook from '$lib/components/icons/IconAddressBook.svelte';
-	import IconShieldCheck from '$lib/components/icons/lucide/IconShieldCheck.svelte';
-	import Button from '$lib/components/ui/Button.svelte';
+	import TransactionsFilterContactsEmptyState from '$lib/components/transactions/filter/TransactionsFilterContactsEmptyState.svelte';
 	import InputSearch from '$lib/components/ui/InputSearch.svelte';
-	import { TRANSACTIONS_FILTER_CONTACTS_EMPTY_CTA } from '$lib/constants/test-ids.constants';
 	import { contacts } from '$lib/derived/contacts.derived';
 	import {
 		PLAUSIBLE_EVENT_EVENTS_KEYS,
@@ -13,10 +10,9 @@
 	} from '$lib/enums/plausible';
 	import { trackTransactionFilter } from '$lib/services/analytics.services';
 	import { i18n } from '$lib/stores/i18n.store';
-	import { modalStore } from '$lib/stores/modal.store';
 	import { transactionsFilterStore } from '$lib/stores/transactions-filter.store';
 	import { matchesContactByText } from '$lib/utils/contact.utils';
-	import { replaceOisyPlaceholders, replacePlaceholders } from '$lib/utils/i18n.utils';
+	import { replacePlaceholders } from '$lib/utils/i18n.utils';
 
 	interface Props {
 		// When the panel is rendered inside a desktop dropdown popover the
@@ -65,45 +61,11 @@
 
 		transactionsFilterStore.toggleContactId(id);
 	};
-
-	const openAddressBook = () => {
-		modalStore.openAddressBook({ id: Symbol() });
-	};
 </script>
 
 <div class="flex flex-col gap-3">
 	{#if isEmpty}
-		<div class="flex flex-col items-center gap-3 px-2 py-4 text-center">
-			<div class="w-16 text-brand-primary [&_svg]:h-auto [&_svg]:w-full" aria-hidden="true">
-				<IconAddressBook />
-			</div>
-
-			<p class="text-sm text-secondary">{$i18n.transaction.filter.contacts_empty_title}</p>
-
-			<p class="text-sm text-tertiary">
-				<strong
-					><span class="relative -top-px mr-1 inline-block align-middle text-success-primary"
-						><IconShieldCheck size="16" /></span
-					>{`${replaceOisyPlaceholders($i18n.core.text.oisy_protects_you)} `}</strong
-				>{$i18n.transaction.filter.contacts_empty_description}<br />
-				<a
-					class="blue-link no-underline"
-					href="https://docs.oisy.com/introduction/oisy-keeps-you-protected#contacts"
-					rel="noopener noreferrer"
-					target="_blank">{$i18n.core.text.learn_more}</a
-				>
-			</p>
-
-			<Button
-				colorStyle="primary"
-				fullWidth
-				onclick={openAddressBook}
-				testId={TRANSACTIONS_FILTER_CONTACTS_EMPTY_CTA}
-				type="button"
-			>
-				{$i18n.transaction.filter.contacts_empty_cta}
-			</Button>
-		</div>
+		<TransactionsFilterContactsEmptyState />
 	{:else}
 		<InputSearch
 			{autofocus}
