+++ src/frontend/src/tests/lib/components/transactions/filter/TransactionsFilterContactsPanel.spec.ts
@@ -1,5 +1,4 @@
 import TransactionsFilterContactsPanel from '$lib/components/transactions/filter/TransactionsFilterContactsPanel.svelte';
-import { TRANSACTIONS_FILTER_CONTACTS_EMPTY_CTA } from '$lib/constants/test-ids.constants';
 import * as contactsDerived from '$lib/derived/contacts.derived';
 import {
 	PLAUSIBLE_EVENT_EVENTS_KEYS,
@@ -7,10 +6,9 @@ import {
 } from '$lib/enums/plausible';
 import * as analyticsServices from '$lib/services/analytics.services';
 import { i18n } from '$lib/stores/i18n.store';
-import { modalStore } from '$lib/stores/modal.store';
 import { transactionsFilterStore } from '$lib/stores/transactions-filter.store';
 import type { ContactUi } from '$lib/types/contact';
-import { replaceOisyPlaceholders, replacePlaceholders } from '$lib/utils/i18n.utils';
+import { replacePlaceholders } from '$lib/utils/i18n.utils';
 import { getMockContactsUi, mockContactEthAddressUi } from '$tests/mocks/contacts.mock';
 import { fireEvent, render } from '@testing-library/svelte';
 import { get } from 'svelte/store';
@@ -134,52 +132,15 @@ describe('TransactionsFilterContactsPanel', () => {
 			).toBeNull();
 		});
 
-		it('renders the no-contacts-yet line above the OISY-protects-you block', () => {
+		// The contents of the empty state itself (icon, OISY lockup, Learn
+		// more link, CTA, etc.) are covered by
+		// `TransactionsFilterContactsEmptyState.spec.ts`. The panel spec just
+		// asserts the branching: an empty contact list switches the panel
+		// from the list view to the empty-state component.
+		it('renders the empty-state component', () => {
 			const { getByText } = render(TransactionsFilterContactsPanel);
 
 			expect(getByText(get(i18n).transaction.filter.contacts_empty_title)).toBeInTheDocument();
-		});
-
-		it('renders the OISY-protects-you lockup and description', () => {
-			const { getByText } = render(TransactionsFilterContactsPanel);
-
-			expect(
-				getByText(replaceOisyPlaceholders(get(i18n).core.text.oisy_protects_you))
-			).toBeInTheDocument();
-			// The description sits between the OISY-protects-you <strong> and the
-			// Learn more <a> inside the same <p>, so the paragraph's full text is
-			// "<oisy lockup> <description> <learn more>". Use a partial match
-			// against the i18n value so the test stays in sync with the copy.
-			expect(
-				getByText(get(i18n).transaction.filter.contacts_empty_description, { exact: false })
-			).toBeInTheDocument();
-		});
-
-		it('renders a Learn more link pointing to the protected-contacts docs', () => {
-			const { getByRole } = render(TransactionsFilterContactsPanel);
-
-			const link = getByRole('link', { name: get(i18n).core.text.learn_more });
-
-			expect(link).toHaveAttribute(
-				'href',
-				'https://docs.oisy.com/introduction/oisy-keeps-you-protected#contacts'
-			);
-			expect(link).toHaveAttribute('target', '_blank');
-			expect(link).toHaveAttribute('rel', 'noopener noreferrer');
-		});
-
-		it('renders the CTA button that opens the address book', async () => {
-			const openSpy = vi.spyOn(modalStore, 'openAddressBook');
-
-			const { getByTestId } = render(TransactionsFilterContactsPanel);
-
-			const cta = getByTestId(TRANSACTIONS_FILTER_CONTACTS_EMPTY_CTA);
-
-			expect(cta).toHaveTextContent(get(i18n).transaction.filter.contacts_empty_cta);
-
-			await fireEvent.click(cta);
-
-			expect(openSpy).toHaveBeenCalledOnce();
 		});
 	});
